Assignment-1 (Submission Doc)
Submit a PDF document containing the project scope and the technical and functional requirements.

You need to define the project's scope in terms of the data you want to store. 
The technical requirements would also be related to the data only, for example the access constraints. 


You need to submit ONLY one document per group containing all of the above. 
Naming Convention: GroupNumber_ProjectTitle_Member1RollNumber